store: separate tests that need Store from those needing *store
